pyproject.toml
[tool.poetry]
name = "gx-snowflake"
version = "0.1.0"
description = ""
authors = ["bru-cab <cabrerabr1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
ipython = "^8.18.1"
ipykernel = "^6.27.1"
python-dotenv = "^1.0.0"
great-expectations = "^0.18.5"
snowflake-connector-python = "^3.6.0"
pip = "^23.3.2"
install = "^1.3.5"
snowflake-snowpark-python = "^1.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
